Added book filetype detection and reorganized tests due to size of file
