+++ API.Tests/Parser/BookParserTests.cs
@@ -1,0 +1,7 @@
+﻿namespace API.Tests
+{
+    public class BookParserTests
+    {
+        
+    }
+}
